refactor: rename mutations
