+++ src/schema/v2/offer/__tests__/rejectSellerOfferMutation.test.ts
@@ -3,7 +3,7 @@ import { baseOrderJson } from "../../order/__tests__/support"
 
 const mockMutation = `
   mutation {
-    buyerAcceptOffer(input: {
+    rejectSellerOffer(input: {
       orderID: "order-id",
       offerID: "offer-id"
     }) {
@@ -12,11 +12,6 @@ const mockMutation = `
           mutationError {
             message
             code
-          }
-        }
-        ...on OrderMutationActionRequired {
-          actionData {
-            clientSecret
           }
         }
         ...on OrderMutationSuccess {
@@ -31,10 +26,10 @@ const mockMutation = `
 
 let context
 
-describe("buyerAcceptOfferMutation", () => {
+describe("rejectSellerOfferMutation", () => {
   beforeEach(() => {
     context = {
-      meOfferAcceptLoader: jest.fn().mockResolvedValue({
+      meOfferRejectLoader: jest.fn().mockResolvedValue({
         ...baseOrderJson,
         id: "order-id",
         mode: "offer",
@@ -42,12 +37,12 @@ describe("buyerAcceptOfferMutation", () => {
     }
   })
 
-  it("accepts a seller's offer and returns the order", async () => {
+  it("declines a seller's offer and returns the order", async () => {
     const result = await runAuthenticatedQuery(mockMutation, context)
 
     expect(result.errors).toBeUndefined()
     expect(result).toEqual({
-      buyerAcceptOffer: {
+      rejectSellerOffer: {
         orderOrError: {
           order: {
             internalID: "order-id",
@@ -56,18 +51,46 @@ describe("buyerAcceptOfferMutation", () => {
       },
     })
 
-    expect(context.meOfferAcceptLoader).toHaveBeenCalledWith({
-      orderID: "order-id",
-      offerID: "offer-id",
-    })
+    expect(context.meOfferRejectLoader).toHaveBeenCalledWith(
+      { orderID: "order-id", offerID: "offer-id" },
+      {}
+    )
+  })
+
+  it("forwards rejectReason as reject_reason", async () => {
+    const mutation = `
+      mutation {
+        rejectSellerOffer(input: {
+          orderID: "order-id",
+          offerID: "offer-id",
+          rejectReason: "buyer_rejected"
+        }) {
+          orderOrError {
+            ...on OrderMutationSuccess {
+              order {
+                internalID
+              }
+            }
+          }
+        }
+      }
+    `
+
+    const result = await runAuthenticatedQuery(mutation, context)
+
+    expect(result.errors).toBeUndefined()
+    expect(context.meOfferRejectLoader).toHaveBeenCalledWith(
+      { orderID: "order-id", offerID: "offer-id" },
+      { reject_reason: "buyer_rejected" }
+    )
   })
 
   it("returns a 422 exchange error", async () => {
-    context.meOfferAcceptLoader = jest.fn().mockRejectedValue({
+    context.meOfferRejectLoader = jest.fn().mockRejectedValue({
       statusCode: 422,
       body: {
-        message: "offer_total_not_defined",
-        code: "offer_total_not_defined",
+        message: "cannot_reject_offer",
+        code: "cannot_reject_offer",
       },
     })
 
@@ -75,34 +98,11 @@ describe("buyerAcceptOfferMutation", () => {
 
     expect(result.errors).toBeUndefined()
     expect(result).toEqual({
-      buyerAcceptOffer: {
+      rejectSellerOffer: {
         orderOrError: {
           mutationError: {
-            message: "offer_total_not_defined",
-            code: "offer_total_not_defined",
-          },
-        },
-      },
-    })
-  })
-
-  it("returns an action-required response when payment requires action", async () => {
-    context.meOfferAcceptLoader = jest.fn().mockRejectedValue({
-      statusCode: 422,
-      body: {
-        code: "payment_requires_action",
-        action_data: { client_secret: "cs_1" },
-      },
-    })
-
-    const result = await runAuthenticatedQuery(mockMutation, context)
-
-    expect(result.errors).toBeUndefined()
-    expect(result).toEqual({
-      buyerAcceptOffer: {
-        orderOrError: {
-          actionData: {
-            clientSecret: "cs_1",
+            message: "cannot_reject_offer",
+            code: "cannot_reject_offer",
           },
         },
       },
@@ -114,7 +114,7 @@ describe("buyerAcceptOfferMutation", () => {
 
     expect(result.errors).toBeUndefined()
     expect(result).toEqual({
-      buyerAcceptOffer: {
+      rejectSellerOffer: {
         orderOrError: {
           mutationError: {
             message: "An error occurred",
