+++ src/schema/v2/offer/acceptSellerOfferMutation.ts
@@ -1,4 +1,4 @@
-import { GraphQLNonNull, GraphQLID, GraphQLString } from "graphql"
+import { GraphQLNonNull, GraphQLID } from "graphql"
 import {
   ORDER_MUTATION_FLAGS,
   OrderMutationResponseType,
@@ -10,16 +10,15 @@ import { handleExchangeError } from "../order/exchangeErrorHandling"
 interface Input {
   orderID: string
   offerID: string
-  rejectReason?: string
 }
 
-export const buyerRejectOfferMutation = mutationWithClientMutationId<
+export const acceptSellerOfferMutation = mutationWithClientMutationId<
   Input,
   any,
   ResolverContext
 >({
-  name: "buyerRejectOffer",
-  description: "Decline a seller's offer on an order",
+  name: "acceptSellerOffer",
+  description: "Accept a seller's offer on an order",
   inputFields: {
     orderID: {
       type: new GraphQLNonNull(GraphQLID),
@@ -27,11 +26,7 @@ export const buyerRejectOfferMutation = mutationWithClientMutationId<
     },
     offerID: {
       type: new GraphQLNonNull(GraphQLID),
-      description: "Offer id to decline.",
-    },
-    rejectReason: {
-      type: GraphQLString,
-      description: "Optional reason for declining the offer.",
+      description: "Offer id to accept.",
     },
   },
   outputFields: {
@@ -41,23 +36,13 @@ export const buyerRejectOfferMutation = mutationWithClientMutationId<
     },
   },
   mutateAndGetPayload: async (input, context) => {
-    const { meOfferRejectLoader } = context
-    if (!meOfferRejectLoader) {
+    const { meOfferAcceptLoader } = context
+    if (!meOfferAcceptLoader) {
       throw new Error("You need to be signed in to perform this action")
     }
     try {
       const { orderID, offerID } = input
-      const exchangeInputFields = {
-        reject_reason: input.rejectReason,
-      }
-
-      const payload = Object.fromEntries(
-        Object.entries(exchangeInputFields).filter(
-          ([_, value]) => value !== undefined
-        )
-      )
-
-      const order = await meOfferRejectLoader({ orderID, offerID }, payload)
+      const order = await meOfferAcceptLoader({ orderID, offerID })
       order._type = ORDER_MUTATION_FLAGS.SUCCESS // Set the type for the response
       return order
     } catch (error) {
