+++ src/schema/v2/offer/rejectSellerOfferMutation.ts
@@ -1,4 +1,4 @@
-import { GraphQLNonNull, GraphQLID } from "graphql"
+import { GraphQLNonNull, GraphQLID, GraphQLString } from "graphql"
 import {
   ORDER_MUTATION_FLAGS,
   OrderMutationResponseType,
@@ -10,15 +10,16 @@ import { handleExchangeError } from "../order/exchangeErrorHandling"
 interface Input {
   orderID: string
   offerID: string
+  rejectReason?: string
 }
 
-export const buyerAcceptOfferMutation = mutationWithClientMutationId<
+export const rejectSellerOfferMutation = mutationWithClientMutationId<
   Input,
   any,
   ResolverContext
 >({
-  name: "buyerAcceptOffer",
-  description: "Accept a seller's offer on an order",
+  name: "rejectSellerOffer",
+  description: "Decline a seller's offer on an order",
   inputFields: {
     orderID: {
       type: new GraphQLNonNull(GraphQLID),
@@ -26,7 +27,11 @@ export const buyerAcceptOfferMutation = mutationWithClientMutationId<
     },
     offerID: {
       type: new GraphQLNonNull(GraphQLID),
-      description: "Offer id to accept.",
+      description: "Offer id to decline.",
+    },
+    rejectReason: {
+      type: GraphQLString,
+      description: "Optional reason for declining the offer.",
     },
   },
   outputFields: {
@@ -36,13 +41,23 @@ export const buyerAcceptOfferMutation = mutationWithClientMutationId<
     },
   },
   mutateAndGetPayload: async (input, context) => {
-    const { meOfferAcceptLoader } = context
-    if (!meOfferAcceptLoader) {
+    const { meOfferRejectLoader } = context
+    if (!meOfferRejectLoader) {
       throw new Error("You need to be signed in to perform this action")
     }
     try {
       const { orderID, offerID } = input
-      const order = await meOfferAcceptLoader({ orderID, offerID })
+      const exchangeInputFields = {
+        reject_reason: input.rejectReason,
+      }
+
+      const payload = Object.fromEntries(
+        Object.entries(exchangeInputFields).filter(
+          ([_, value]) => value !== undefined
+        )
+      )
+
+      const order = await meOfferRejectLoader({ orderID, offerID }, payload)
       order._type = ORDER_MUTATION_FLAGS.SUCCESS // Set the type for the response
       return order
     } catch (error) {
